solve Qs

+++ 662. Maximum Width of Binary Tree/code.cpp
@@ -12,16 +12,16 @@
 class Solution {
 public:
     int widthOfBinaryTree(TreeNode* root) {
-        queue<pair<TreeNode*, int>> q;
+        queue<pair<TreeNode*, unsigned long long>> q;
         q.push({root, 0});
         int maxWidth = 0;
 
         while(q.size()>0){
             int currLevelSize = q.size();
-            int stIdx = q.front().second;
-            int endIdx = q.back().second;
+            unsigned long long stIdx = q.front().second;
+            unsigned long long endIdx = q.back().second;
             
-            maxWidth = max(maxWidth, endIdx-stIdx+1);
+            maxWidth = max(maxWidth, (int)(endIdx-stIdx+1));
 
             for(int i=0; i<currLevelSize; i++){
                 auto curr = q.front();
